refactor - add const.py, change order of propertys, methods
