+++ homeassistant/components/evohome/const.py
@@ -1,0 +1,9 @@
+"""Provides the constants needed for evohome."""
+
+DOMAIN = 'evohome'
+DATA_EVOHOME = 'data_' + DOMAIN
+DISPATCHER_EVOHOME = 'dispatcher_' + DOMAIN
+
+# These are used only to help prevent E501 (line too long) violations.
+GWS = 'gateways'
+TCS = 'temperatureControlSystems'
